i2c: Add GPIO-based hotplug gate
Implement driver for hot-plugged I2C busses, where some devices on
a bus are hot-pluggable and their presence is indicated by GPIO line.
This feature is used by the ASUS Transformers family, by the  Microsoft
Surface RT/2 and that may be not the full list.

The Transformers have a connector that's used for USB, charging or for
attaching a dock-keyboard (which also has a battery and a touchpad).
This connector has an I2C bus lines and a "detect" line (pulled low
on the dock side) among the pins. Close look over motherboards of
tablets and docks has shown  no additional chip or a transparent
bridge/buffer chip, nothing that could be controlled by software.

Surface RT/2 have a similar approach apart that the tCover controller
is on separate board inside the tablet but becomes active only if tCover
is attached.

For DT this setup could be modelled like an I2C gate or 2-port mux with
enable joining two I2C buses (one "closer" to the CPU as a parent).

Co-developed-by: Ion Agorria <ion@agorria.com>
Signed-off-by: Ion Agorria <ion@agorria.com>
Signed-off-by: Michał Mirosław <mirq-linux@rere.qmqm.pl>
Signed-off-by: Svyatoslav Ryhel <clamor95@gmail.com>